pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-patterns"
version = "0.1.0"
description = "Multi-agent patterns for LangChain/LangGraph 1.2.x"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "langchain>=1.2.4",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "langgraph-supervisor>=0.0.31",
    "langgraph-swarm>=0.0.15",
    "deepagents>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentic_patterns"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/skills",
    "/examples",
]
